+++ examples/candles.js
@@ -1,8 +1,8 @@
-import { TradingViewClient } from "tvdatafeedclient-js";
+import { TvDataFeed } from "tvdatafeedclient-js";
 
 (async () => {
-  const tv = new TradingViewClient();
-  await tv.connect();
+  const tv = new TvDataFeed();
+  await tv.connect(); //optional
 
   const candles = await tv.getCandles("BINANCE", "BTCUSDT", "1", 300);
   console.table(candles);